[feat/#59] Review 기능 확장

+++ backend/src/main/java/waruru/backend/review/service/ReviewService.java
@@ -36,14 +36,18 @@ public class ReviewService {
 
     @Transactional
     public ReviewResponseDTO createReview(ReviewRequestDTO rrq){
-        Review review = new Review();
 
-        //  예외처리 해주기
+        boolean reviewExists = reviewRepository.existsByUserNo_IdAndSaleNo_No(rrq.getUserNo(), rrq.getSaleNo());
+        if (reviewExists) {
+            throw new IllegalStateException("해당 매물에 관해 작성한 후기가 존재합니다.");
+        }
+
         Member member = memberRepository.findById(rrq.getUserNo())
                 .orElseThrow(() -> new EntityNotFoundException("회원 ID 조회 불가: " + rrq.getUserNo()));
         Sale sale = saleRepository.findById(rrq.getSaleNo())
                 .orElseThrow(() -> new EntityNotFoundException("매물 번호 조회 불가: " + rrq.getSaleNo()));
 
+        Review review = new Review();
         review.setUserNo(member);
         review.setSaleNo(sale);
         review.setTitle(rrq.getTitle());
@@ -75,13 +79,14 @@ public class ReviewService {
         Review review = reviewRepository.findById(reviewNo)
                 .orElseThrow(() -> new EntityNotFoundException("매물 후기 수정 불가: " + reviewNo));
 
+        if (!review.getUserNo().getId().equals(reviewUpdateRequestDTO.getUserNo())) {
+            throw new SecurityException("후기를 수정할 권한이 없습니다.");
+        }
         review.setTitle(reviewUpdateRequestDTO.getTitle());
         review.setContent(reviewUpdateRequestDTO.getContent());
         review.setUpdateDate(reviewUpdateRequestDTO.getUpdateDate());
 
         Review updatedReview = reviewRepository.save(review);
-
-        //feat: 나중에 리뷰작성한 유저인지 확인 후 수정하게끔 로직 구현하기
 
         return new ReviewResponseDTO(updatedReview);
     }
@@ -89,10 +94,13 @@ public class ReviewService {
 
     // 매물 후기 삭제
     @Transactional
-    public void deleteReview(ReviewDeleteRequestDTO deleteReviewRequestDTO) {
-        Review review = reviewRepository.findById(deleteReviewRequestDTO.getReviewNo())
-                .orElseThrow(() -> new EntityNotFoundException("매물 후기 삭제 불가: " + deleteReviewRequestDTO.getReviewNo()));
+    public void deleteReview(Long reviewNo, ReviewDeleteRequestDTO deleteReviewRequestDTO) {
+        Review review = reviewRepository.findById(reviewNo)
+                .orElseThrow(() -> new EntityNotFoundException("매물 후기 삭제 불가: " + reviewNo));
 
+        if (!review.getUserNo().getId().equals(deleteReviewRequestDTO.getUserNo())) {
+            throw new SecurityException("후기를 삭제할 권한이 없습니다.");
+        }
         reviewRepository.delete(review);
     }
